requirements.txt
flask==3.1.0
python-dotenv==1.1.0
openai==1.75.0
pytest==8.3.5
pytest-cov==4.1.0
black==24.3.0
flake8==7.0.0 